- 게시판 수정 - 삭제 버튼 시 confirm - 에디터 태그 제외 후 표시 - 목록 보기 시 수정하기 버튼 추가 - 게시글 상단 고정 - 조회수 늘리기 - 말머리관리 삭제 - 검색기능 완료

+++ src/main/resources/static/js/board/header.js
@@ -54,28 +54,32 @@ function deleteSelectedBoards(boardIds) {
         });
 }
 
-
 function deleteSelectedBoard(boardId, redirectUrl) {
-    fetch('/masterpage_sys/board/api/delete', {
-        method: 'DELETE',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(boardId) // 객체 형태로 감싸서 전송
-    })
-        .then(response => {
-            if (!response.ok) {
-                throw new Error('삭제 실패');
-            }
-            return response.json();
+    // 사용자 확인
+    if (confirm('정말 삭제하시겠습니까?')) {
+        fetch('/masterpage_sys/board/api/delete', {
+            method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify(boardId) // 객체 형태로 감싸서 전송
         })
-        .then(data => {
-            alert(data.response);
-            window.location.href = redirectUrl; // 성공 후 목록으로 이동
-        })
-        .catch(error => {
-            alert('오류가 발생했습니다: ' + error.message);
-            console.error('Error:', error);
-        });
+            .then(response => {
+                if (!response.ok) {
+                    throw new Error('삭제 실패');
+                }
+                return response.json();
+            })
+            .then(data => {
+                alert(data.response); // 성공 메시지 표시
+                window.location.href = redirectUrl; // 성공 후 목록으로 이동
+            })
+            .catch(error => {
+                alert('오류가 발생했습니다: ' + error.message);
+                console.error('Error:', error);
+            });
+    } else {
+        alert('삭제가 취소되었습니다.'); // 사용자가 취소한 경우
+    }
 }
